// src/modules/core/mail/enums/mail-subject.enum.ts
export enum MailSubjectEnum {
  VERIFY_USER_EMAIL = 'VERIFY_USER_EMAIL',
  RESET_USER_PASSWORD = 'RESET_USER_PASSWORD',
  LOGIN_USER_WITH_VERIFICATION_CODE = 'LOGIN_USER_WITH_VERIFICATION_CODE',
  UPDATE_EMAIL_VERIFICATION_CODE = 'UPDATE_EMAIL_VERIFICATION_CODE',
  VENDOR_APPROVAL = 'VENDOR_APPROVAL',
  VENDOR_REJECTION = 'VENDOR_REJECTION',
}