// src/modules/core/mail/enums/mail-template.enum.ts
export enum MailTemplateEnum {
  Welcome = 'Welcome.hbs',
  VERIFY_USER_EMAIL = 'verify-email.hbs',
  RESET_USER_PASSWORD = 'reset-password.hbs',
  LOGIN_USER_WITH_VERIFICATION_CODE = 'login-user-with-verification-code.hbs',
  UPDATE_EMAIL_VERIFICATION_CODE = 'update-email-verification-code.hbs',
  VENDOR_APPROVAL = 'vendor-approval.hbs',
  VENDOR_REJECTION = 'vendor-rejection.hbs',
}